backend/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
python-dotenv==1.0.0
PyJWT==2.6.0
Werkzeug==2.0.1
gunicorn==20.1.0